Move ProvideConnectionPayloadTypes
