+++ crates/relayer/relayer-components/src/chain/impls/types/mod.rs
@@ -1,2 +1,3 @@
 pub mod commitment_prefix;
+pub mod connection_payload;
 pub mod proof;
